Fixed double to int
git-svn-id: https://vmtk.svn.sourceforge.net/svnroot/vmtk@181 94be75cf-be22-0410-a3f1-cbcec1c01a80

+++ vtkVmtk/Segmentation/vtkvmtkVesselEnhancingDiffusion3DImageFilter.h
@@ -101,7 +101,7 @@ class VTK_VMTK_SEGMENTATION_EXPORT vtkvmtkVesselEnhancingDiffusion3DImageFilter 
     DelegateITKInputMacro(SetGamma,value);
   }
 
-  void SetRecalculateVesselness(double value)
+  void SetRecalculateVesselness(int value)
   {
     DelegateITKInputMacro(SetRecalculateVesselness,value);
   }
